referral_system/requirements.txt
asgiref                       3.8.1
Django                        5.1.2
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
drf-yasg                      1.21.8
inflection                    0.5.1
mysqlclient                   2.2.5
packaging                     24.1
pip                           24.3.1
psycopg                       3.2.3
PyJWT                         2.9.0
python-decouple               3.8
pytz                          2024.2
PyYAML                        6.0.2
sqlparse                      0.5.1
tzdata                        2024.2
uritemplate                   4.1.1